requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.6
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
defusedxml==0.7.1
Django==2.2.24
django-colorfield==0.4.5
django-filter==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
drf-base64==2.0
flake8==4.0.1
gunicorn==20.1.0
idna==3.3
isort==5.10.0
itypes==1.2.0
Jinja2==3.0.2
MarkupSafe==2.0.1
mccabe==0.6.1
oauthlib==3.1.1
Pillow==8.4.0
psycopg2-binary==2.8.6
py==1.10.0
pycodestyle==2.8.0
pycparser==2.20
pyflakes==2.4.0
PyJWT==2.1.0
python-dotenv==0.19.1
python-slugify==5.0.2
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sorl-thumbnail==12.7.0
sqlparse==0.4.2
testfixtures==6.18.3
text-unidecode==1.3
uritemplate==4.1.1
urllib3==1.26.7
